Better error handling
Also used the occasion to update dependencies and move away from fs-promise
in favor of fs-extra. This is also the first commit to add package-lock.json
to the repo.

This commit mainly handles when no matching bundle for a requested file can
be found and also removes some error-callbacks in promises which would discard
the error.